requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.5
dj-database-url==0.5.0
Django==3.2.6
django-mathfilters==1.0.0
gunicorn==20.1.0
idna==3.2
numpy==1.21.2
pandas==1.3.3
pycodestyle==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
Unipath==1.1
urllib3==1.26.6
whitenoise==5.3.0
django-heroku
openpyxl
